Syscalls now also work with 0x80 interrupt (and syscall instruction partially)

+++ src/aero_kernel/src/syscall/process.rs
@@ -1,6 +1,6 @@
 use crate::userland::scheduler;
 
-pub fn exit(status: i32) {
+pub fn exit(status: usize) {
     log::debug!("Exiting the current process with status: {}", status);
 
     scheduler::get_scheduler();
